Moved all Icons to one folder, to export from components
