+++ src/lib/components/EditNote/editNote.helpers.ts
@@ -1,5 +1,5 @@
 export const generateCopy = (formId: string | undefined, title: string) => ({
 	header: formId ? `Edit ${title}` : 'Doot a new note 📯',
 	buttonText: formId ? 'Save changes' : 'Create note',
-	delayedReason: formId ? 'Updating note' : 'Creating note',
+	submitDelayedReason: formId ? 'Updating note' : 'Creating note',
 })
