--- src/lib/components/icons/README.md
@@ -1,5 +1,0 @@
-# Icons
-
-This directory contains SVG icons that are used by the app.
-
-Everything in this directory is generated by `npm run build:icons`.
