+++ src/lib/components/icons/figureOutSpriteIcons/README.md
@@ -1,0 +1,5 @@
+# Icons
+
+This directory contains SVG icons that are used by the app.
+
+Everything in this directory is generated by `npm run build:icons`.
