+++ src/lib/components/icons/figureOutSpriteIcons/name.d.ts
@@ -1,0 +1,3 @@
+// This file is generated by npm run build:icons
+
+export type IconName = 'magnifying-glass'
