+++ src/lib/components/icons/index.ts
@@ -1,0 +1,39 @@
+import Timer from 'virtual:icons/radix-icons/timer'
+import Eraser from 'virtual:icons/radix-icons/eraser'
+import Pencil1 from 'virtual:icons/radix-icons/pencil1'
+import Pencil2 from 'virtual:icons/radix-icons/pencil2'
+import Person from 'virtual:icons/radix-icons/person'
+import Exit from 'virtual:icons/radix-icons/exit'
+import Camera from 'virtual:icons/radix-icons/camera'
+import DotsHorizontal from 'virtual:icons/radix-icons/dotsHorizontal'
+import EnvelopeClosed from 'virtual:icons/radix-icons/envelopeClosed'
+import LockClosed from 'virtual:icons/radix-icons/lockClosed'
+import LockOpened from 'virtual:icons/radix-icons/lock-open-1'
+import Download from 'virtual:icons/radix-icons/download'
+import Link2 from 'virtual:icons/radix-icons/link-2'
+import Check from 'virtual:icons/radix-icons/check'
+import Cross from 'virtual:icons/radix-icons/cross2'
+import Plus from 'virtual:icons/radix-icons/plus'
+import MagnifyingGlass from 'virtual:icons/radix-icons/magnifying-glass'
+import Trash from 'virtual:icons/radix-icons/trash'
+
+export {
+	Timer,
+	Eraser,
+	Pencil1,
+	Pencil2,
+	Person,
+	Exit,
+	Camera,
+	DotsHorizontal,
+	EnvelopeClosed,
+	LockClosed,
+	LockOpened,
+	Download,
+	Link2,
+	Check,
+	Cross,
+	Plus,
+	MagnifyingGlass,
+	Trash,
+}
