--- src/lib/components/icons/name.d.ts
@@ -1,3 +1,0 @@
-// This file is generated by npm run build:icons
-
-export type IconName = 'magnifying-glass'
