+++ src/lib/components/index.ts
@@ -2,7 +2,8 @@ export { default as AvatarMenu } from '$lib/components/AvatarMenu/AvatarMenu.sve
 export { default as Button } from '$lib/components/Button/Button.svelte'
 export { default as EditNote } from '$lib/components/EditNote/EditNote.svelte'
 export { default as HoneypotInputs } from '$lib/components/HoneypotInputs/HoneypotInputs.svelte'
-export { default as Icon } from '$lib/components/icons/Icon.svelte'
+// export { default as Icon } from '$lib/components/icons/figureOutSpriteIcons/Icon.svelte'
+export * from '$lib/components/icons'
 export { default as ImageEditor } from '$lib/components/ImageEditor/ImageEditor.svelte'
 export { Input, TextArea, FormGroup } from '$lib/components/FormElements'
 export { default as Navbar } from '$lib/components/Navbar/Navbar.svelte'
