+++ src/routes/users/[username]/notes/+layout.svelte
@@ -1,14 +1,12 @@
 <script lang="ts">
 	import { page } from '$app/stores'
-	import { Button } from '$lib/components'
-	import Pencil1 from 'virtual:icons/radix-icons/pencil1'
+	import { Button, Pencil1 } from '$lib/components'
 
 	export let data
 	const hrefBase = `/users/${$page.params.username}/notes`
 	$: isOwner = $page.data?.user?.id === data.owner.id
 </script>
 
-<!-- TODO: Find out why the notes layout isn't rendering 🙄 -->
 <div class="wrapper" data-layout="grid">
 	<div class="sidebar">
 		<h1>{data.owner.name}'s notes</h1>
