+++ src/routes/users/[username]/notes/[noteId]/+page.server.ts
@@ -15,7 +15,6 @@ export const actions = {
 			},
 			where: { id: params.noteId, ownerId: userId },
 		})
-		await new Promise(fulfil => setTimeout(fulfil, 3000))
 		invariantResponse(note, 'Not found', 404)
 		const isOwner = note.ownerId === userId
 		// TODO: Check if user has permissions
